+++ src/objects.js
@@ -87,7 +87,7 @@ BlockVisibilityDialogMorph*/
 
 /*jshint esversion: 6*/
 
-modules.objects = '2021-October-14';
+modules.objects = '2021-October-21';
 
 var SpriteMorph;
 var StageMorph;
@@ -751,16 +751,6 @@ SpriteMorph.prototype.initBlocks = function () {
             category: 'control',
             spec: 'when %b'
         },
-        doBroadcast: {
-            type: 'command',
-            category: 'control',
-            spec: 'broadcast %msg'
-        },
-        doBroadcastAndWait: {
-            type: 'command',
-            category: 'control',
-            spec: 'broadcast %msg and wait'
-        },
         getLastMessage: {  // retained for legacy compatibility
             dev: true,
             type: 'reporter',
@@ -770,7 +760,14 @@ SpriteMorph.prototype.initBlocks = function () {
         doSend: {
             type: 'command',
             category: 'control',
-            spec: 'send %msg to %spr'
+            spec: 'send %msg to %rcv',
+            defaults: [null, ['all']]
+        },
+        doSendAndWait: {
+            type: 'command',
+            category: 'control',
+            spec: 'send %msg to %rcv and wait',
+            defaults: [null, ['all']]
         },
         doWait: {
             type: 'command',
@@ -907,11 +904,6 @@ SpriteMorph.prototype.initBlocks = function () {
         },
 
         // Scenes
-        receiveOnScene: {
-            type: 'hat',
-            category: 'control',
-            spec: 'when switched to this scene %message'
-        },
         doSwitchToScene: {
             type: 'command',
             category: 'control',
@@ -1659,6 +1651,16 @@ SpriteMorph.prototype.initBlockMigrations = function () {
             selector: 'reportListAttribute',
             inputs: [['length']],
             offset: 1
+        },
+        doBroadcast: {
+            selector: 'doSend',
+            inputs: [null, ['all']],
+            offset: 0
+        },
+        doBroadcastAndWait: {
+            selector: 'doSendAndWait',
+            inputs: [null, ['all']],
+            offset: 0
         }
     };
 };
@@ -1730,9 +1732,8 @@ SpriteMorph.prototype.blockAlternatives = {
     setSize: ['changeSize'],
 
     // control:
-    doBroadcast: ['doBroadcastAndWait', 'doSend'],
-    doBroadcastAndWait: ['doBroadcast', 'doSend'],
-    doSend: ['doBroadcast', 'doBroadcastAndWait'],
+    doSend: ['doSendAndWait'],
+    doSendAndWait: ['doSend'],
     doIf: ['doIfElse', 'doUntil'],
     doIfElse: ['doIf', 'doUntil'],
     doRepeat: ['doUntil', ['doForever', -1], ['doFor', 2], ['doForEach', 1]],
@@ -2521,9 +2522,8 @@ SpriteMorph.prototype.blockTemplates = function (
         blocks.push(block('receiveCondition'));
         blocks.push('-');
         blocks.push(block('receiveMessage'));
-        blocks.push(block('doBroadcast'));
-        blocks.push(block('doBroadcastAndWait'));
         blocks.push(block('doSend'));
+        blocks.push(block('doSendAndWait'));
         blocks.push('-');
         blocks.push(block('doWarp'));
         blocks.push('-');
@@ -2557,10 +2557,8 @@ SpriteMorph.prototype.blockTemplates = function (
         blocks.push(block('newClone'));
         blocks.push(block('removeClone'));
         blocks.push('-');
-        blocks.push(block('receiveOnScene'));
-        blocks.push(block('doSwitchToScene'));
-        blocks.push('-');
         blocks.push(block('doPauseAll'));
+        blocks.push(block('doSwitchToScene'));
 
         // for debugging: ///////////////
         if (devMode) {
@@ -6095,17 +6093,13 @@ SpriteMorph.prototype.allHatBlocksFor = function (message) {
                 return event === message
                     || (event instanceof Array
                         && message !== '__shout__go__'
-                        && message !== '__clone__init__'
-                        && message !== '__scene__init__');
+                        && message !== '__clone__init__');
             }
             if (sel === 'receiveGo') {
                 return message === '__shout__go__';
             }
             if (sel === 'receiveOnClone') {
                 return message === '__clone__init__';
-            }
-            if (sel === 'receiveOnScene') {
-                return message === '__scene__init__';
             }
         }
         return false;
@@ -8472,24 +8466,33 @@ StageMorph.prototype.fireChangeOfSceneEvent = function (message) {
 
     this.children.concat(this).forEach(morph => {
         if (isSnapObject(morph)) {
-            morph.allHatBlocksFor('__scene__init__').forEach(block => {
-                var varName =  block.inputs()[0].evaluate()[0],
-                    varFrame;
-                if (varName) {
-                    varFrame = new VariableFrame();
-                    varFrame.addVar(varName, isNil(message) ? '' : message);
+            morph.allHatBlocksFor(message).forEach(block => {
+                var varName, varFrame;
+                if (block.selector === 'receiveMessage') {
+                    varName = block.inputs()[1].evaluate()[0];
+                    if (varName) {
+                        varFrame = new VariableFrame();
+                        varFrame.addVar(varName, message);
+                    }
+                    procs.push(this.threads.startProcess(
+                        block,
+                        morph,
+                        this.isThreadSafe || // make "any msg" threadsafe
+                            block.inputs()[0].evaluate() instanceof Array,
+                        null, // exportResult (bool)
+                        null, // callback
+                        null, // isClicked
+                        null, // rightAway
+                        null, // atomic
+                        varFrame
+                    ));
+                } else {
+                    procs.push(this.threads.startProcess(
+                        block,
+                        morph,
+                        this.isThreadSafe
+                    ));
                 }
-                procs.push(this.threads.startProcess(
-                    block,
-                    morph,
-                    this.isThreadSafe,
-                    null, // exportResult (bool)
-                    null, // callback
-                    null, // isClicked
-                    null, // rightAway
-                    null, // atomic
-                    varFrame
-                ));
             });
         }
     });
@@ -8765,9 +8768,8 @@ StageMorph.prototype.blockTemplates = function (
         blocks.push(block('receiveCondition'));
         blocks.push('-');
         blocks.push(block('receiveMessage'));
-        blocks.push(block('doBroadcast'));
-        blocks.push(block('doBroadcastAndWait'));
         blocks.push(block('doSend'));
+        blocks.push(block('doSendAndWait'));
         blocks.push('-');
         blocks.push(block('doWarp'));
         blocks.push('-');
@@ -8799,10 +8801,8 @@ StageMorph.prototype.blockTemplates = function (
         blocks.push(block('createClone'));
         blocks.push(block('newClone'));
         blocks.push('-');
-        blocks.push(block('receiveOnScene'));
-        blocks.push(block('doSwitchToScene'));
-        blocks.push('-');
         blocks.push(block('doPauseAll'));
+        blocks.push(block('doSwitchToScene'));
 
         // for debugging: ///////////////
         if (this.world().isDevMode) {